Do not send null input name attributes on xhr send
Prevent adding formData to XHR with a null key. This happens when the
input attribute on a form is unset and causes an error.

I needed to do this because the submit button was being added to the
xhr formData (Amazon S3 can restrict requests to specific header and
this wasn't specified) and removing the name attribute from the submit
button sent a 'null' key in post.