Allow for removing pre-defined request headers
Sometimes, we don't want to send those pre-defined headers along with the request. This is really useful when dealing with CORS requests: if we call setRequestHeader("headername", "some value") method, then the preflight OPTIONS request (related to CORS) will have it's header "Access-Control-Request-Headers" referencing the headername header. So, this means for the CORS request to work, the response of the OPTIONS request must have the XX header listed in it's "Access-Control-Allow-Headers". In other words, if we prompt access to "headername" header (via Acces-Control-Request-Headers), then we must be explicitely allowed to use it (via Access-Control-Allow-Headers). And as often with CORS, we might not be able to modify the server response because we're using a third party website, so letting the user the ability to remove it from the initial request is a workaround to do CORS request properly.