Replace "acceptParameter" with "acceptedFiles"
Dropzone.js does not actually have any "acceptParameter" as parameter, but only "acceptedFiles" and the deprecated "acceptedMimeTypes" parameter. In this test suite some tests refer to "acceptedParameter" which would be okay if this is understood as a summary of the two aforementioned possible parameters.

However in some tests "acceptedParameter" were actually provided as initialization parameter to Dropzone instead of "acceptedFiles" which makes no sense whatsoever (dropzone.coffee does not refer to "acceptedParameter").

To prevent any future confusion I also removed any reference to "acceptedParameter" in the test descriptions in favour of more precise descriptions.