Adds boolean parameter 'sendFileAsBody'
This will change the outgoing request to not send a form at all and instead send the file as the body of the request. This is useful for web servers which will itself upload the file to another source (for example S3), so they can more easily forward the file content while still receiving the request. This will show the user a more accurate progress.